Move `TypeMappingImpl` into `MappingBuilder`
